change 'Accepted' -> Godkan (required to get tests to pass)
The factories have to do something different if building an 'Accepted/Godkan' membership.  Because we currently compare the hardcoded string value for the status, and so because the factory is checking for the string "Godkan", needed to change the status from "Accepted" to "Godkan" so that the membership status is set appropriately and the company is associated correctly.