refactor; validation: include test for lat, long; complete geocoding
after_validation needs to call :geocode_best_possible if either latitude or longitude or both are missing. Geocoding: needed more complete tests for #geocode_all_needed, and for

Needed to add logic to _not_ geocode if we are doing development and testing and already have latitude & longitude.  Needed so we can read in an address from CSV (with lat & long) and _not_ have it be geocoded.