use milliseconds to create unique email
preview will actually call each method twice: once for the header and once for the body. If it happens fast enough, it will happen in less than 1 second, and so the 2 email strings created ("new_email" done twice) won't be unique.  Thus we have to use milliseconds to create a unique email name.