make more detailed step for finding info on a row with Co. name
Cucumber is having trouble parsing the steps and coming up with the step we indend because we have too many steps that aren't quite specific enough.  Once we have the new design in place we can ensure that we have sufficient ids and class tags so that our steps will be unambiguous.